+++ src/AppserverIo/Appserver/ServletEngine/RequestHandler.php
@@ -128,11 +128,11 @@ class RequestHandler extends \Thread
             }
 
         } catch (\Exception $e) {
-            error_log($e->__toString());
-            $servletResponse->appendBodyStream($e->__toString());
-            $servletResponse->setStatusCode(500);
+            // bind the exception in the respsonse
+            $servletResponse->setException($e);
         }
 
+        // shutdown the request handler
         $this->__shutdown();
     }
 
@@ -168,6 +168,7 @@ class RequestHandler extends \Thread
             $servletResponse->appendBodyStream($lastError['message']);
         }
 
+        // shutdown the request handler
         $this->__shutdown();
     }
 }